refactor: Refactor project organization
