+++ user/src/main/java/com/rentit/user/controller/AuthController.java
@@ -2,8 +2,7 @@ package com.rentit.user.controller;
 
 import com.rentit.user.dto.UserLoginDTO;
 import com.rentit.user.dto.UserRegistrationDTO;
-import com.rentit.user.service.AuthService;
-import com.rentit.user.service.UserService;
+import com.rentit.user.service.OnboardService;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -15,21 +14,19 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/api/users")
-public class UserController {
+public class AuthController {
     @Autowired
-    private UserService userService;
-    @Autowired
-    private AuthService authService;
+    private OnboardService onboardService;
 
     @PostMapping("/register")
     public ResponseEntity<String> registerUser(@Valid @RequestBody UserRegistrationDTO userRegistrationDTO){
-        userService.registerUser(userRegistrationDTO);
+        onboardService.registerUser(userRegistrationDTO);
         return ResponseEntity.status(HttpStatus.OK).body("User has been registered successfully");
     }
 
     @PostMapping("/login")
     public ResponseEntity<String> loginUser(@Valid @RequestBody UserLoginDTO userLoginDTO){
-        return authService.login(userLoginDTO);
+        return onboardService.loginUser(userLoginDTO);
     }
 
 }
